edit: organizacion de carpetas y archivos
